app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.3.0
aiosmtplib==5.0.0
python-dotenv==1.0.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dateutil==2.9.0.post0
httpx==0.27.0